back/requirements.txt
uvicorn==0.23.1
fastapi==0.100.0
python-socketio==5.8.0
eventlet==0.33.3
python-dotenv==1.0.0